Forgot to add gamepadcontroller script. my bad Added useItem function to gamepad & KeyboardMouse controller Scripts tweaked the following prefabs: oil slick - added "oil slick" tag, changed shape to plane, tweaked positions during instantiation where it will always be on the ground (unable to move it away from player using it causing them to slip, needs more work). freezeshot & movement script - resized prefab, changed projectile speed to 20, changed initial position to be 3 units further from vehicle firing (in some cases, any ai player using freeze shot will cause player to lose their freeze shot powerup item. unable to confirm), added "Freeze" tag
Tweaked the following scripts:
Vehicle - added the UseItem() method for powerups and added vehicleItem script components
FreezeShotMovement & vehicleItems scripts had very minor changes to reflect the changes to prefabs above
AIDriver & AIState - added condition to "freeze" their controls when frozen by freeze shot.
StatsGUI - Added Player Item to the list (but for some reason, it doesn't show on the screen).
PlayerUI - Managed to fill the bar with the texture (but has this weird way of extending beyond the bottom of the screen, can be fixed though).